move types to utils
