+++ packages/controllers/package.json
@@ -61,7 +61,6 @@
     "@metamask/eslint-config-jest": "^9.0.0",
     "@metamask/eslint-config-nodejs": "^9.0.0",
     "@metamask/eslint-config-typescript": "^9.0.1",
-    "@metamask/snap-types": "^0.19.0",
     "@metamask/template-snap": "^0.7.0",
     "@peculiar/webcrypto": "^1.3.3",
     "@types/concat-stream": "^2.0.0",
