+++ packages/controllers/src/services/node/NodeProcessExecutionService.test.ts
@@ -1,5 +1,5 @@
 import { ControllerMessenger } from '@metamask/controllers';
-import { ErrorJSON, SnapId } from '@metamask/snap-types';
+import { ErrorJSON, SnapId } from '@metamask/snap-utils';
 import { JsonRpcEngine } from 'json-rpc-engine';
 import { createEngineStream } from 'json-rpc-middleware-stream';
 import pump from 'pump';
