+++ packages/examples/examples/browserify/package.json
@@ -27,7 +27,7 @@
     "@metamask/eslint-config-jest": "^9.0.0",
     "@metamask/eslint-config-nodejs": "^9.0.0",
     "@metamask/eslint-config-typescript": "^9.0.1",
-    "@metamask/snap-types": "^0.19.0",
+    "@metamask/snap-utils": "^0.19.0",
     "@metamask/snaps-browserify-plugin": "^0.19.0",
     "@metamask/snaps-cli": "^0.19.0",
     "@types/browserify": "^12.0.37",
