+++ packages/examples/examples/browserify/tsconfig.json
@@ -3,6 +3,6 @@
   "compilerOptions": {
     "resolveJsonModule": true
   },
-  "files": ["../../../types/global.d.ts"],
+  "files": ["../../../utils/global.d.ts"],
   "include": ["src", "gulpfile.ts", "babel.config.json"]
 }
