+++ packages/examples/examples/rollup/package.json
@@ -30,7 +30,7 @@
     "@metamask/eslint-config-nodejs": "^9.0.0",
     "@metamask/eslint-config-typescript": "^9.0.1",
     "@metamask/rollup-plugin-snaps": "^0.19.0",
-    "@metamask/snap-types": "^0.19.0",
+    "@metamask/snap-utils": "^0.19.0",
     "@metamask/snaps-cli": "^0.19.0",
     "@rollup/plugin-babel": "^5.3.1",
     "@types/node": "^17.0.33",
