+++ packages/examples/examples/typescript/README.md
@@ -9,6 +9,6 @@ This Snap demonstrates how to develop a Snap with TypeScript.
 - For the global `wallet` type to work, you have to add the following to your `tsconfig.json`:
   ```json
   {
-    "files": ["./node_modules/@metamask/snap-types/global.d.ts"]
+    "files": ["./node_modules/@metamask/snap-utils/global.d.ts"]
   }
   ```
