+++ packages/examples/examples/typescript/package.json
@@ -32,7 +32,7 @@
     "@metamask/eslint-config-jest": "^9.0.0",
     "@metamask/eslint-config-nodejs": "^9.0.0",
     "@metamask/eslint-config-typescript": "^9.0.1",
-    "@metamask/snap-types": "^0.19.0",
+    "@metamask/snap-utils": "^0.19.0",
     "@metamask/snaps-cli": "^0.19.0",
     "eslint": "^7.30.0",
     "eslint-config-prettier": "^8.3.0",
