+++ packages/examples/examples/typescript/src/index.ts
@@ -1,4 +1,4 @@
-import { OnRpcRequestHandler } from '@metamask/snap-types';
+import { OnRpcRequestHandler } from '@metamask/snap-utils';
 import { getMessage } from './message';
 
 /**
