+++ packages/examples/examples/webpack/package.json
@@ -29,7 +29,7 @@
     "@metamask/eslint-config-jest": "^9.0.0",
     "@metamask/eslint-config-nodejs": "^9.0.0",
     "@metamask/eslint-config-typescript": "^9.0.1",
-    "@metamask/snap-types": "^0.19.0",
+    "@metamask/snap-utils": "^0.19.0",
     "@metamask/snaps-cli": "^0.19.0",
     "@metamask/snaps-webpack-plugin": "^0.19.0",
     "@types/node": "^17.0.33",
