+++ packages/examples/examples/webpack/tsconfig.json
@@ -3,6 +3,6 @@
   "compilerOptions": {
     "typeRoots": ["../../../../node_modules/@types"]
   },
-  "files": ["../../../types/global.d.ts"],
+  "files": ["../../../utils/global.d.ts"],
   "include": ["src", "webpack.config.ts"]
 }
