+++ packages/execution-environments/package.json
@@ -35,7 +35,7 @@
     "@metamask/object-multiplex": "^1.2.0",
     "@metamask/post-message-stream": "^6.0.0",
     "@metamask/providers": "^9.0.0",
-    "@metamask/snap-types": "^0.19.0",
+    "@metamask/snap-utils": "^0.19.0",
     "@metamask/utils": "^2.0.0",
     "eth-rpc-errors": "^4.0.3",
     "pump": "^3.0.0",
