--- packages/types/README.md
@@ -1,3 +1,0 @@
-# @metamask/snap-types
-
-Types for MetaMask Snaps packages.
