--- packages/types/package.json
@@ -1,54 +1,0 @@
-{
-  "name": "@metamask/snap-types",
-  "version": "0.19.0",
-  "description": "Shared types for MetaMask Snaps packages.",
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/MetaMask/snaps-skunkworks.git"
-  },
-  "main": "src/index.d.ts",
-  "types": "src/index.d.ts",
-  "files": [
-    "src",
-    "global.d.ts"
-  ],
-  "scripts": {
-    "lint:eslint": "eslint . --cache --ext js,ts",
-    "lint:misc": "prettier '**/*.json' '**/*.md' '!CHANGELOG.md' --ignore-path ../../.gitignore",
-    "lint": "yarn lint:eslint && yarn lint:misc --check",
-    "lint:fix": "yarn lint:eslint --fix && yarn lint:misc --write",
-    "lint:changelog": "yarn auto-changelog validate",
-    "publish:package": "../../scripts/publish-package.sh"
-  },
-  "dependencies": {
-    "@metamask/controllers": "^30.0.0"
-  },
-  "devDependencies": {
-    "@lavamoat/allow-scripts": "^2.0.3",
-    "@metamask/auto-changelog": "^2.6.0",
-    "@metamask/eslint-config": "^9.0.0",
-    "@metamask/eslint-config-jest": "^9.0.0",
-    "@metamask/eslint-config-nodejs": "^9.0.0",
-    "@metamask/eslint-config-typescript": "^9.0.1",
-    "@metamask/providers": "^9.0.0",
-    "@metamask/types": "^1.1.0",
-    "eslint": "^7.30.0",
-    "eslint-config-prettier": "^8.3.0",
-    "eslint-plugin-import": "^2.23.4",
-    "eslint-plugin-jest": "^24.4.0",
-    "eslint-plugin-jsdoc": "^36.1.0",
-    "eslint-plugin-node": "^11.1.0",
-    "eslint-plugin-prettier": "^3.4.0",
-    "prettier": "^2.3.2",
-    "prettier-plugin-packagejson": "^2.2.11",
-    "rimraf": "^3.0.2",
-    "typescript": "^4.4.0"
-  },
-  "engines": {
-    "node": ">=16.0.0"
-  },
-  "publishConfig": {
-    "access": "public",
-    "registry": "https://registry.npmjs.org/"
-  }
-}
