--- packages/types/src/index.d.ts
@@ -1,1 +1,0 @@
-export * from './types';
