+++ packages/utils/package.json
@@ -7,7 +7,8 @@
   },
   "main": "dist/index.js",
   "files": [
-    "dist/"
+    "dist/",
+    "global.d.ts"
   ],
   "scripts": {
     "test": "jest && yarn posttest",
@@ -29,6 +30,9 @@
   },
   "dependencies": {
     "@babel/core": "^7.18.6",
+    "@metamask/controllers": "^30.3.0",
+    "@metamask/providers": "^9.1.0",
+    "@metamask/types": "^1.1.0",
     "@metamask/utils": "^2.0.0",
     "ajv": "^8.11.0",
     "fast-deep-equal": "^3.1.3",
