+++ packages/utils/src/types.ts
@@ -1,4 +1,6 @@
 import { Json } from '@metamask/utils';
+import { MetaMaskInpageProvider } from '@metamask/providers';
+import { JsonRpcRequest } from '@metamask/types';
 import { NpmSnapPackageJson, SnapManifest } from './json-schemas';
 
 /**
@@ -42,3 +44,38 @@ export enum SnapValidationFailureReason {
   RepositoryMismatch = '"repository" field mismatch',
   ShasumMismatch = '"shasum" field mismatch',
 }
+
+/**
+ * Command request sent to a worker.
+ */
+export type WorkerCommandRequest = {
+  id: string;
+  command: string;
+  data?: string | Record<string, unknown>;
+};
+
+export type SnapData = {
+  snapId: string;
+  sourceCode: string;
+};
+
+export type SnapExecutionData = SnapData & {
+  endowments?: Json;
+};
+
+export type SnapRpcHandler = (args: {
+  origin: string;
+  request: JsonRpcRequest<unknown[] | { [key: string]: unknown }>;
+}) => Promise<unknown>;
+
+export type OnRpcRequestHandler = SnapRpcHandler;
+
+export type SnapProvider = MetaMaskInpageProvider;
+
+export type SnapId = string;
+
+export type ErrorJSON = {
+  message: string;
+  code: number;
+  data?: Json;
+};
